requirement.txt
SpeechRecognition
python-dotenv
openai 
pipwin
pipwin install pyaudio


